allow to download full zip updates
states are now reflected, like downloading, downloaded.

TODO: installed

i am using a threadpoolexecuter and workqueues, which allow to download
multiple updates at once, if someone wants to.

while downloading a notification is present, for each download.
pressing on these notifications cause the update list to open, where a
download can be aborted if wanted.

downloaded updates can be deleted or installed.

before installing we check the md5sum of the downloaded file with the
provided md5sum to be sure the download is not corrupt

if the md5sum does not match, show a warning, else let the user reboot
and install.

NOTE: the notification bypasses this check

Change-Id: I00d83606d20f0582ac530285ad8900fe35a70fbe