allow to display changelog
introduce another button bar at the expanded content.
depending on the buttons available, it hides itself.

it checks for the changelog in the files directory and if found, loads it.
else it loads it from sourceforge and puts it in the files directory to
cache it for the next time.

Change-Id: Ief29c227c1be6910f9637a1046d07853ffc16cbb