do not use a broadcast for checking updates, change bus un-/registration
register in onStart, unregister in onStop.
this is to be sure we only register once and only unregister if the
fragment got destroyed.

else it could happen that the download finished succeeded but the UI did
not update and we still show "downloading" and the user has to reopen the
screen.

Change-Id: Ic872055b52b7bf9fe1d75c496c2a13e84f256db4